requirements.txt
decorator==4.0.10
Django==1.10.2
ipdb==0.10.1
ipython==5.1.0
ipython-genutils==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==3.4.2
prompt-toolkit==1.0.8
ptyprocess==0.5.1
Pygments==2.1.3
pytesseract==0.1.6
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
wcwidth==0.1.7
